+++ swarm/CHANGELOG.md
@@ -18,6 +18,11 @@
   This variant was never constructed and thus dead code.
   See [PR 3605].
 
+- Remove deprecated `IntoConnectionHandler` and all its implementations.
+  This also removes the `NetworkBehaviour::new_handler` and `NetworkBehaviour::addresses_of_peer` methods.
+  See changelog for `0.42` on how to migrate.
+  See [PR 3884].
+
 - Remove `ConnectionHandlerUpgrErr::Timer` variant.
   This variant was never constructed and thus dead code.
   See [PR 3605].
@@ -31,6 +36,9 @@
 [PR 3746]: https://github.com/libp2p/rust-libp2p/pull/3746
 [PR 3865]: https://github.com/libp2p/rust-libp2p/pull/3865
 [PR 3886]: https://github.com/libp2p/rust-libp2p/pull/3886
+[PR 3884]: https://github.com/libp2p/rust-libp2p/pull/3884
+[PR 3605]: https://github.com/libp2p/rust-libp2p/pull/3605
+[PR 3746]: https://github.com/libp2p/rust-libp2p/pull/3746
 [PR 3882]: https://github.com/libp2p/rust-libp2p/pull/3882
 [PR 3605]: https://github.com/libp2p/rust-libp2p/pull/3605
 [PR 3746]: https://github.com/libp2p/rust-libp2p/pull/3746